♻ Move function out of utils module.
